fix issue of yaxis truncated, fix issue indicator description, and counter down disappear issue
