+++ sass/components/studylegend.scss
@@ -316,11 +316,11 @@
         height: 100%;
         padding: 16px 24px;
         box-sizing: border-box;
-        z-index: 1;
         display: flex;
         flex-direction: column;
         align-items: center;
         justify-content: space-between;
+        z-index: 9999;
         @include themify($themes) {
             color: themed('IndicatorInfoText');
             background: themed('IndicatorInfoBg');
