+++ src/chartiq_injections/index.js
@@ -6,6 +6,8 @@ import { maintainSpanSize }       from './resizing';
 import { setMaxTicks }            from './setMaxTicks';
 import { plotterDrawText }        from './plotterDrawText';
 import { overideMeasure }         from './setMeasure';
+import { registerHTMLElements }   from './registerHTMLElements';
+import { stickyInterior }         from './stickyInterior';
 
 const inject = (option) => {
     createXAxis();
@@ -16,6 +18,8 @@ const inject = (option) => {
     plotterDrawText();
     setMaxTicks();
     overideMeasure(option);
+    registerHTMLElements();
+    stickyInterior(option);
 };
 
 export default inject;
